--- weeek2/style.css
@@ -1,112 +1,0 @@
-html {
-  font-size: 16px;
-}
-
-header {
-  width: 100%;
-  text-align: center;
-  font-weight: bold;
-  background-color: rgb(230, 230, 230);
-  padding: 10px;
-  border-bottom: 1px solid gray;
-}
-
-nav {
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-  margin: 10px 0;
-}
-
-nav > button {
-  background-color: rgb(230, 230, 230);
-  padding: 5px 20px;
-  font-weight: bold;
-  font-size: 14px;
-  border: none;
-  cursor: pointer;
-}
-
-main {
-  height: calc(100vh - 104px);
-  display: flex;
-  background-color: rgb(230, 230, 230);
-  border-top: 1px solid gray;
-}
-
-main > section {
-  flex: 1;
-  text-align: center;
-  padding: 20px;
-}
-
-main > section:first-child {
-  border-right: 1px solid gray;
-}
-
-section h1 {
-  text-align: center;
-}
-
-.todo-wrapper {
-  height: calc(100vh - 184px);
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-ul {
-  margin: 20px 0;
-  padding: 5px 10px;
-  max-height: 500px;
-  overflow-y: scroll;
-}
-
-ul li {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px dashed gray;
-  padding: 2px;
-  margin: 10px 0;
-}
-
-ul li > div {
-  text-align: left;
-}
-
-article {
-  width: 100%;
-  border: none;
-  border-bottom: 1px solid gray;
-  background-color: white;
-  padding: 4px 10px;
-  display: flex;
-  justify-content: space-between;
-  gap: 10px;
-}
-
-article > button {
-  border: none;
-}
-
-input {
-  width: 100%;
-  border: none;
-  outline: none;
-}
-
-footer {
-  width: 100%;
-  position: fixed;
-  bottom: 0;
-  padding: 3px 0;
-  border-top: 1px solid gray;
-  font-weight: bold;
-  text-align: center;
-}
-
-.hidden {
-  width: 0;
-  display: none;
-}
