+++ week2/todo/index.html
@@ -14,52 +14,45 @@
     <script defer src="./index.js"></script>
   </head>
   <body>
-    <header>
-      <h1>TO DO list</h1>
-    </header>
+    <header>TO DO List</header>
 
-    <div class="nav">
-      <button class="nav-goToday">View to day</button>
-      <button class="nav-goTomorrow">View tomorrow</button>
-      <button class="go-togeter">View together</button>
-    </div>
+    <nav>
+      <button id="view-today">View to day</button>
+      <button id="view-tomorrow">View tomorrow</button>
+      <button id="view-together">View together</button>
+    </nav>
 
     <main>
-      <div class="left">
-        <h1>오늘 할일</h1>
-        <ul id="list-today">
-          <li>
-            <span> do 과제 끝내기.</span
-            ><button class="today-delete">
-              <i class="fa-regular fa-trash-can"></i>
-            </button>
-          </li>
-        </ul>
-        <div class="text">
-          <input type="text" id="input-today" /><button id="add-today">
-            <i class="fa-sharp fa-solid fa-plus"></i>
-          </button>
-        </div>
-      </div>
+      <section class="left">
+        <h1>오늘 할 일</h1>
+        <div class="todo-wrapper">
+          <ul id="list-today"></ul>
 
-      <div class="right">
-        <h1>내일 할일</h1>
-        <ul id="list-tomorrow">
-          <li>
-            <span>velog 과제 끝내기.</span
-            ><button><i class="fa-regular fa-trash-can"></i></button>
-          </li>
-        </ul>
-        <div class="text">
-          <input type="text" id="input-tomorrow" /><button id="add-tomorrow">
-            <i class="fa-sharp fa-solid fa-plus"></i>
-          </button>
+          <article>
+            <input type="text" id="input-today" />
+            <button id="add-today">
+              <i class="fa-sharp fa-solid fa-plus"></i>
+            </button>
+          </article>
         </div>
-      </div>
+      </section>
+
+      <section class="right">
+        <h1>내일 할 일</h1>
+
+        <div class="todo-wrapper">
+          <ul id="list-tomorrow"></ul>
+
+          <article>
+            <input type="text" id="input-tomorrow" />
+            <button id="add-tomorrow">
+              <i class="fa-sharp fa-solid fa-plus"></i>
+            </button>
+          </article>
+        </div>
+      </section>
     </main>
 
-    <footer>
-      <h2>copyrightⓒhyun wook</h2>
-    </footer>
+    <footer>copyrightⓒhyun wook</footer>
   </body>
 </html>
