+++ week2/todo/style.css
@@ -1,127 +1,112 @@
-* {
-  font-weight: bold;
+html {
+  font-size: 16px;
 }
+
 header {
-  border-bottom: 2px solid gray;
+  width: 100%;
+  text-align: center;
+  font-weight: bold;
   background-color: rgb(230, 230, 230);
-  height: 70px;
-  text-align: center;
-  display: block;
+  padding: 10px;
+  border-bottom: 1px solid gray;
 }
-body {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-header h1 {
-  font-weight: bold;
-  font-size: 30px;
-  padding-top: 18px;
-}
-.nav > button {
-  border: none;
-  width: 200px;
-  color: rgb(0, 0, 0);
-  font-size: 18px;
-  font-weight: bold;
-  text-align: center;
-  transition: 0.5s all;
-  overflow: hidden;
-}
-.nav > button:hover {
-  transform: scale(1.1);
-}
-.nav > button:after {
-  content: "";
-  position: absolute;
-  top: calc(50% - 50px);
-  left: -50px;
-  width: 20px;
-  height: 100px;
-  background: white;
-  -webkit-transform: rotate(30deg);
-  transform: rotate(30deg);
-  -webkit-transition: all 1000ms cubic-bezier(0.23, 1, 0.32, 1);
-  transition: all 1000ms cubic-bezier(0.23, 1, 0.32, 1);
-  opacity: 0;
-}
-.nav > button:hover:after {
-  left: calc(100% + 50px);
-  opacity: 1;
-}
-.nav {
-  border-bottom: 2px grey solid;
-  height: 100px;
+
+nav {
   display: flex;
   justify-content: center;
-  line-height: 3.1;
+  gap: 10px;
+  margin: 10px 0;
 }
-.nav button {
-  margin: 30px 10px;
-  height: 30px;
+
+nav > button {
   background-color: rgb(230, 230, 230);
+  padding: 5px 20px;
+  font-weight: bold;
+  font-size: 14px;
+  border: none;
+  cursor: pointer;
 }
+
 main {
+  height: calc(100vh - 104px);
   display: flex;
-  justify-content: space-around;
-  box-sizing: border-box;
-  background-color: rgb(218, 218, 218);
-  height: 713px;
+  background-color: rgb(230, 230, 230);
+  border-top: 1px solid gray;
 }
-main div {
-  width: 500px;
-  border: 1px solid gray;
+
+main > section {
+  flex: 1;
+  text-align: center;
+  padding: 20px;
+}
+
+main > section:first-child {
+  border-right: 1px solid gray;
+}
+
+section h1 {
+  text-align: center;
+}
+
+.todo-wrapper {
+  height: calc(100vh - 184px);
   display: flex;
   flex-direction: column;
-  align-items: center;
-}
-main .left > ul > li,
-main .right > ul > li {
-  margin: 40px;
-  border-bottom: 1px gray dashed;
-  width: 400px;
-  padding-bottom: 7px;
-  display: flex;
   justify-content: space-between;
 }
 
-.right > h1,
-.left > h1 {
-  font-size: 2rem;
+ul {
+  margin: 20px 0;
+  padding: 5px 10px;
+  max-height: 500px;
+  overflow-y: scroll;
 }
 
-main .left > .text,
-main .right > .text {
-  position: absolute;
-  bottom: 0;
-  height: 30px;
-  width: 400px;
-  border: none;
-  border-bottom: 2px gray solid;
-  outline: none;
+ul li {
   display: flex;
-  justify-content: center;
+  justify-content: space-between;
+  align-items: center;
+  border-bottom: 1px dashed gray;
+  padding: 2px;
+  margin: 10px 0;
 }
-#add-tomorrow,
-#add-today {
-  position: relative;
-  top: 200px;
-  left: 30px;
+
+ul li > div {
+  text-align: left;
 }
-footer {
-  position: absolute;
-  left: 0;
-  bottom: 0;
+
+article {
   width: 100%;
-  height: 40px;
+  border: none;
+  border-bottom: 1px solid gray;
+  background-color: white;
+  padding: 4px 10px;
+  display: flex;
+  justify-content: space-between;
+  gap: 10px;
+}
+
+article > button {
+  border: none;
+}
+
+input {
+  width: 100%;
+  border: none;
+  outline: none;
+}
+
+footer {
+  width: 100%;
+  position: fixed;
+  bottom: 0;
+  padding: 3px 0;
+  border-top: 1px solid gray;
+  font-weight: bold;
   text-align: center;
-  border-top: 2px gray solid;
 }
-footer h2 {
-  margin-top: 10px;
-}
+
 .hidden {
   width: 0;
-  overflow: hidden;
   display: none;
 }
