+++ week2/velog/index.html
@@ -1,0 +1,204 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="style.css" />
+    <script
+      defer
+      src="https://kit.fontawesome.com/e327aa4fae.js"
+      crossorigin="anonymous"
+    ></script>
+    <script src="./script.js"></script>
+    <title>WOOKLOG</title>
+  </head>
+  <body>
+    <header>
+      <div class="head-left">
+        <h1>WOOKLOG</h1>
+      </div>
+      <section class="head-right">
+        <span><i class="fa-solid fa-magnifying-glass"></i></span>
+        <span><strong>새 글 작성</strong></span>
+        <span><img src="./img/dragon.png" alt="리자몽" /></span>
+        <span><i class="fa-solid fa-arrow-down"></i></span>
+      </section>
+    </header>
+    <nav>
+      <section class="nav-left">
+        <button>
+          <span><i class="fa-solid fa-arrow-trend-up"></i></span>트렌딩
+        </button>
+        <button>
+          <span><i class="fa-regular fa-clock"></i></span>최신
+        </button>
+        <button>
+          이번 주<span><i class="fa-solid fa-arrow-down"></i></span>
+        </button>
+      </section>
+      <section class="nav-right">
+        <button class="dot">
+          <i class="fa-solid fa-ellipsis-vertical"></i>
+        </button>
+      </section>
+    </nav>
+    <main>
+      <section>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+        <article>
+          <img src="./img/wep.jpg" alt="이게 솝트다!!!!" />
+          <h2>IN SOPT</h2>
+          <ul>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+            <li><strong>3줄이야~~</strong></li>
+          </ul>
+          <span>5일 전 35개의 댓글</span>
+          <div>
+            <span
+              ><img src="./img/dragon.png" alt="리자몽" />by
+              <strong> wook_did</strong></span
+            >
+            <span><i class="fa-sharp fa-solid fa-heart"></i>96</span>
+          </div>
+        </article>
+      </section>
+    </main>
+  </body>
+</html>
