+++ week2/velog/style.css
@@ -1,0 +1,116 @@
+@font-face {
+  font-family: "MonoplexKR-Italic";
+  src: url("https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_Monoplex-kr@1.0/MonoplexKR-Italic.woff2")
+    format("woff2");
+  font-weight: 400;
+  font-style: normal;
+}
+body {
+  margin: 0px auto;
+  padding: 0px;
+  width: 1024px;
+  box-sizing: border-box;
+  font-family: "twayair";
+  background-color: #e9ecef;
+}
+header {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+}
+header .head-right span {
+  margin: 0px 0.8rem;
+}
+header .head-right > span:nth-child(2) {
+  border-radius: 30px;
+  border: 1px solid black;
+  padding: 3px 8px;
+}
+header .head-right img {
+  border-radius: 50%;
+  width: 40px;
+  position: relative;
+  top: 0.7rem;
+}
+nav button:hover {
+  font-size: 150%;
+  border-bottom: 2px solid black;
+  color: black;
+}
+nav {
+  display: flex;
+  justify-content: space-between;
+}
+nav span {
+  padding: 0px 0.8rem;
+}
+nav button {
+  border: none;
+  background-color: #e9ecef;
+  font-size: 120%;
+  margin: 0px 0.7rem;
+  color: gray;
+}
+nav .nav-left button:nth-child(1) {
+  border-bottom: 3px solid black;
+  padding-bottom: 8px;
+  color: black;
+}
+main {
+  width: 100%;
+  height: 100%;
+}
+main section {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  flex-wrap: wrap;
+}
+main article {
+  background-color: #ffffff;
+  margin: 1.5rem 0.5rem;
+  width: 30%;
+}
+article img {
+  width: 100%;
+}
+article > h2 {
+  padding: 0.2rem 0.8rem;
+  margin: 0;
+  font-size: 1.3rem;
+}
+article ul {
+  padding: 0;
+  margin: 0.6rem 0.7rem;
+}
+article ul > li {
+  list-style: none;
+  font-size: 1rem;
+  margin: 0.3rem 0;
+}
+article > span {
+  font-size: 0.8rem;
+  color: gray;
+  display: flex;
+  margin: 0rem 0rem 1rem 0.7rem;
+}
+article > div {
+  width: 100%;
+  height: 3rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+article div > span > img {
+  border-radius: 50%;
+  width: 25px;
+  margin-right: 1rem;
+}
+article div > span:nth-child(1) {
+  margin-left: 1rem;
+  line-height: 3rem;
+  display: flex;
+}
+article div > span:nth-child(2) {
+  margin-right: 1rem;
+}
