+++ week3/src/App.jsx
@@ -1,0 +1,13 @@
+import Header from "./components/Header";
+import Nav from "./components/Nav";
+
+function App() {
+  return (
+    <div className="App">
+      <Header />
+      <Nav />
+    </div>
+  );
+}
+
+export default App;
