+++ week3/src/components/Header.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import styled from "styled-components";
+export default function Header() {
+  return <H1>안녕하세요.</H1>;
+}
+
+const H1 = styled.h1`
+  color: red;
+`;
