+++ week3/src/components/Nav.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Nav() {
+  return <div>네비게이야</div>;
+}
